Renaming, moving and refactoring auto-retry functionality
